image-to-text/consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Emyr298 <emi.ru298@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
pydantic = "^2.10.4"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
